refactor(trade): 重构价格计算系统，引入计算器模式
- 新增促销活动价格计算器模块 (calculators/)
  * bargain_calculator - 砍价活动计算器
  * combination_calculator - 组合套餐计算器
  * coupon_calculator - 优惠券计算器
  * delivery_calculator - 运费计算器
  * discount_activity_calculator - 折扣活动计算器
  * point_activity_calculator - 积分活动计算器
  * point_give_calculator - 积分赠送计算器
  * point_use_calculator - 积分抵扣计算器
  * reward_activity_calculator - 满减活动计算器
  * seckill_calculator - 秒杀活动计算器
- 删除旧的 price.go 单体文件
- 新增 price_calculator.go 价格计算器协调器
- 新增 price_calculator_helper.go 辅助工具
- 新增 price_service.go 价格服务层
- 新增 types.go 类型定义
- 新增 errors.go 错误定义
- 新增 service_interfaces.go 接口定义

采用策略模式 + 计算器模式，提升代码可维护性和扩展性。

+++ internal/service/product/spu.go
@@ -169,7 +169,7 @@ func (s *ProductSpuService) UpdateSpuStatus(ctx context.Context, req *req.Produc
 // UpdateBrowseCount 更新浏览量
 func (s *ProductSpuService) UpdateBrowseCount(ctx context.Context, id int64, count int) error {
 	_, err := s.q.ProductSpu.WithContext(ctx).Where(s.q.ProductSpu.ID.Eq(id)).
-		Update(s.q.ProductSpu.BrowseCount, s.q.ProductSpu.BrowseCount.Add(int32(count)))
+		Update(s.q.ProductSpu.BrowseCount, s.q.ProductSpu.BrowseCount.Add(count))
 	return err
 }
 
@@ -373,7 +373,7 @@ func (s *ProductSpuService) UpdateSpuStock(ctx context.Context, stockIncr map[in
 		// Note: We don't strictly check SPU stock >= 0 here because it's an aggregate.
 		// SKU level check is the authority.
 		_, err := s.q.ProductSpu.WithContext(ctx).Where(s.q.ProductSpu.ID.Eq(spuID)).
-			Update(s.q.ProductSpu.Stock, s.q.ProductSpu.Stock.Add(int32(incr)))
+			Update(s.q.ProductSpu.Stock, s.q.ProductSpu.Stock.Add(int(incr)))
 		if err != nil {
 			return err
 		}
